# docs/api.yaml
openapi: 3.0.3
...
info:
  title: OpenTogetherTube REST API
  description: This specifies all the endpoints available on OpenTogetherTube. All
    endpoints are subject to change spontaneously and without warning as new
    features are added or bugs are fixed.
  version: 0.2.0
...
servers:
  - url: http://localhost:8080/
  - url: https://staging.opentogethertube.com/
  - url: https://opentogethertube.com/
...
paths:
...
  "/api/room/{RoomName}":
    parameters:
      - $ref: "#/components/parameters/RoomName"
    get:
      tags:
        - room
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      grants:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
                      autoSkipSegments:
                        type: boolean
                      users:
                        type: array
                        items:
                          type: object
                          properties:
                            status:
                              type: string
                            role:
                              type: number
                            name:
                              type: string
                            isLoggedIn:
                              type: boolean
                            id:
                              type: string
                          required:
                            - status
                            - role
                            - name
                            - isLoggedIn
                            - id
                      visibility:
                        type: string
                      title:
                        type: string
                      queueMode:
                        type: string
                      queue:
                        type: array
                        items: {}
                      permissions:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
                      name:
                        type: string
                      isTemporary:
                        type: boolean
                      hasOwner:
                        type: boolean
                      description:
                        type: string
                      clients:
                        type: array
                        items:
                          type: object
                          properties:
                            username:
                              type: string
                            name:
                              type: string
                            isLoggedIn:
                              type: boolean
                          required:
                            - isLoggedIn
                    required:
                      - grants
                      - autoSkipSegments
                      - users
                      - visibility
                      - title
                      - queueMode
                      - queue
                      - name
                      - isTemporary
                      - hasOwner
                      - description
                      - clients
                  - type: array
                    items: {}
        "304":
          description: ""
          content: {}
    patch:
      tags:
        - room
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                autoSkipSegments:
                  type: boolean
                claim:
                  type: boolean
                visibility:
                  type: string
                title:
                  type: string
                queueMode:
                  type: string
                permissions:
                  type: object
                  properties:
                    "0":
                      type: number
                    "1":
                      type: number
                    "2":
                      type: number
                    "3":
                      type: number
                    "4":
                      type: number
                    "-1":
                      type: number
                  required:
                    - "4"
                    - "3"
                    - "2"
                    - "1"
                    - "0"
                  description: ""
                description:
                  type: string
              required:
                - autoSkipSegments
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "500":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: object
                    properties:
                      name:
                        type: string
                      message:
                        type: string
                    required:
                      - name
                      - message
                required:
                  - success
                  - error
    delete:
      description: Unloads the room from memory, if present. Does **not** delete
        permanent rooms from the database.
      tags:
        - room
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
  /api/room/create:
    post:
      tags:
        - room
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  type: string
                title:
                  type: string
                temporary:
                  type: boolean
                queueMode:
                  type: string
                name:
                  type: string
                description:
                  type: string
              required:
                - visibility
                - title
                - temporary
                - queueMode
                - name
                - description
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
components:
  parameters:
    RoomName:
      name: RoomName
      in: path
      schema:
        type: string
      required: true
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
    apikeyAuth:
      type: apiKey
      in: header
      name: apikey